Correct the labels on plots
It is a common nuisance of not having a 3D surface for one of the
mutants (mut #28, sahH). In fact, it could be resolved by fitting the
monotherapy dose-responses manually and taking parameters from there.
This I did for the other surface analysis. For now, this is a low
priority issue for mutants.